Move unrelated packages out of compiler dir
