+++ internal/diagnostics/diagnostics.go
