+++ internal/diagnostics/diagnostics_generated.go
