+++ internal/diagnostics/generate.go
