+++ internal/diagnostics/stringer_generated.go
