+++ internal/tsoptions/commandlineoption.go
@@ -2,8 +2,8 @@ package tsoptions
 
 import (
 	"github.com/microsoft/typescript-go/internal/collections"
-	"github.com/microsoft/typescript-go/internal/compiler/diagnostics"
 	"github.com/microsoft/typescript-go/internal/core"
+	"github.com/microsoft/typescript-go/internal/diagnostics"
 )
 
 type CommandLineOptionKind string
