+++ internal/tsoptions/declscompiler.go
@@ -3,8 +3,8 @@ package tsoptions
 import (
 	"slices"
 
-	"github.com/microsoft/typescript-go/internal/compiler/diagnostics"
 	"github.com/microsoft/typescript-go/internal/core"
+	"github.com/microsoft/typescript-go/internal/diagnostics"
 )
 
 var OptionsDeclarations = slices.Concat(commonOptionsWithBuild, optionsForCompiler)
