+++ internal/tsoptions/declswatch.go
@@ -1,8 +1,8 @@
 package tsoptions
 
 import (
-	"github.com/microsoft/typescript-go/internal/compiler/diagnostics"
 	"github.com/microsoft/typescript-go/internal/core"
+	"github.com/microsoft/typescript-go/internal/diagnostics"
 )
 
 var optionsForWatch = []*CommandLineOption{
